# Optimize Endpoint and Output