--- a/core/src/main/scala/io/finch/Output.scala
+++ b/core/src/main/scala/io/finch/Output.scala
@@ -10,124 +10,121 @@ import java.nio.charset.{Charset, StandardCharsets}
  */
 sealed trait Output[+A] { self =>
 
-  protected def meta: Output.Meta
-  protected def withMeta(meta: Output.Meta): Output[A]
-
   /**
    * The status code of this [[Output]].
    */
-  def status: Status = meta.status
+  def status: Status
 
   /**
    * The header map of this [[Output]].
    */
-  def headers: Map[String, String] = meta.headers
+  def headers: Map[String, String]
 
   /**
    * The cookie list of this [[Output]].
    */
-  def cookies: Seq[Cookie] = meta.cookies
+  def cookies: Seq[Cookie]
 
   /**
    * The charset of this [[Output]].
    */
-  def charset: Option[Charset] = meta.charset
+  def charset: Option[Charset]
 
   /**
    * Returns the payload value of this [[Output]] or throws an exception.
    */
   def value: A
 
-  def map[B](fn: A => B): Output[B] = this match {
-    case Output.Payload(v, m) => Output.Payload(fn(v), m)
-    case f @ Output.Failure(_, _) => f
-    case e @ Output.Empty(_) => e
+  final def map[B](fn: A => B): Output[B] = this match {
+    case p: Output.Payload[A] => p.withValue(fn(p.value))
+    case f: Output.Failure => f
+    case e: Output.Empty => e
   }
 
-  def flatMap[B](fn: A => Output[B]): Output[B] = this match {
-    case p @ Output.Payload(v, _) =>
-      val ob = fn(v)
-      ob.withMeta(ob.meta.copy(
-        headers = ob.headers ++ p.headers,
-        cookies = ob.cookies ++ p.cookies
-      ))
-
-    case f @ Output.Failure(_, _) => f
-    case e @ Output.Empty(_) => e
+  final def flatMap[B](fn: A => Output[B]): Output[B] = this match {
+    case p: Output.Payload[A] => fn(p.value).withCookies(p.cookies).withHeaders(p.headers)
+    case f: Output.Failure => f
+    case e: Output.Empty => e
   }
 
-  def flatten[B](implicit ev: A <:< Output[B]): Output[B] = this match {
-    case Output.Payload(v, _) => v
-    case f @ Output.Failure(_, _) => f
-    case e @ Output.Empty(_) => e
+  final def traverse[B](fn: A => Future[B]): Future[Output[B]] = this match {
+    case p: Output.Payload[A] => fn(p.value).map(b => p.withValue(b))
+    case f: Output.Failure => Future.value(f)
+    case e: Output.Empty => Future.value(e)
   }
 
-  def traverse[B](fn: A => Future[B]): Future[Output[B]] = this match {
-    case p @ Output.Payload(v, _) => fn(v).map(b => p.copy(value = b))
-    case f @ Output.Failure(_, _) => Future.value(f)
-    case e @ Output.Empty(_) => Future.value(e)
+  final def traverseFlatten[B](fn: A => Future[Output[B]]): Future[Output[B]] = this match {
+    case p: Output.Payload[A] =>
+      fn(p.value).map(ob => ob.withHeaders(self.headers).withCookies(self.cookies))
+    case f: Output.Failure => Future.value(f)
+    case e: Output.Empty => Future.value(e)
   }
 
   /**
    * Overrides `charset` of this [[Output]].
    */
-  def withCharset(charset: Charset): Output[A] =
-    withMeta(meta.copy(charset = Some(charset)))
+  final def withCharset(charset: Charset): Output[A] =
+    copy(charset = Some(charset))
 
   /**
    * Overrides the `status` code of this [[Output]].
    */
-  def withStatus(status: Status): Output[A] =
-    withMeta(meta.copy(status = status))
+  final def withStatus(status: Status): Output[A] =
+    copy(status = status)
+
+  /**
+   * Adds given `headers` to this [[Output]].
+   */
+  final def withHeaders(headers: Map[String, String]): Output[A] =
+    if (headers.isEmpty) this
+    else copy(headers = self.headers ++ headers)
+
+  /**
+   * Adds given `cookies` to this [[Output]].
+   */
+  final def withCookies(cookies: Seq[Cookie]): Output[A] =
+    if (cookies.isEmpty) this
+    else copy(cookies = self.cookies ++ cookies)
 
   /**
    * Adds a given `header` to this [[Output]].
    */
-  def withHeader(header: (String, String)): Output[A] =
-    withMeta(meta.copy(headers = headers + header))
+  final def withHeader(header: (String, String)): Output[A] = withHeaders(Map(header))
 
   /**
    * Adds a given `cookie` to this [[Output]].
    */
-  def withCookie(cookie: Cookie): Output[A] =
-    withMeta(meta.copy(cookies = cookies :+ cookie))
+  final def withCookie(cookie: Cookie): Output[A] = withCookies(Seq(cookie))
+
+  protected def copy(status: Status = self.status,
+    charset: Option[Charset] = self.charset,
+    headers: Map[String, String] = self.headers,
+    cookies: Seq[Cookie] = self.cookies): Output[A]
 }
 
 object Output {
 
   /**
-   * A data type representing an HTTP response metadata shared between different types of
-   * [[Output]]s.
-   */
-  private[finch] case class Meta(
-    status: Status = Status.Ok,
-    charset: Option[Charset] = Option.empty,
-    headers: Map[String, String] = Map.empty[String, String],
-    cookies: Seq[Cookie] = Seq.empty[Cookie]
-  )
-
-  /**
    * Creates a successful [[Output]] that wraps a payload `value` with given `status`.
    */
   final def payload[A](value: A, status: Status = Status.Ok): Output[A] =
-    Payload(value, Meta(status = status))
+    Payload(value, status)
 
   /**
    * Creates a failure [[Output]] that wraps an exception `cause` causing this.
    */
   final def failure[A](cause: Exception, status: Status = Status.BadRequest): Output[A] =
-    Failure(cause, Meta(status = status))
+    Failure(cause, status)
 
   /**
    * Creates an empty [[Output]] of given `status`.
    */
-  final def empty[A](status: Status): Output[A] =
-    Empty(Meta(status = status))
+  final def empty[A](status: Status): Output[A] = Empty(status)
 
   /**
    * Creates a unit/empty [[Output]] (i.e., `Output[Unit]`) of given `status`.
    */
-  final def unit(status: Status): Output[Unit] = Empty(Meta(status = status))
+  final def unit(status: Status): Output[Unit] = empty(status)
 
   /**
    * An [[Output]] with `None` as payload.
@@ -137,33 +134,61 @@ object Output {
   /**
    * A successful [[Output]] that captures a payload `value`.
    */
-  private[finch] final case class Payload[A](value: A, meta: Meta) extends Output[A] {
-    override protected def withMeta(meta: Meta): Output[A] = copy(meta = meta)
+  private[finch] final case class Payload[A](
+      value: A,
+      status: Status = Status.Ok,
+      charset: Option[Charset] = Option.empty,
+      headers: Map[String, String] = Map.empty[String, String],
+      cookies: Seq[Cookie] = Seq.empty[Cookie]) extends  Output[A] { self =>
+
+    def withValue[B](value: B): Payload[B] = Payload(value, status, charset, headers, cookies)
+
+    protected def copy(
+        status: Status,
+        charset: Option[Charset],
+        headers: Map[String, String],
+        cookies: Seq[Cookie]): Output[A] = Payload(value, status, charset, headers, cookies)
   }
 
   /**
    * A failure [[Output]] that captures an  [[Exception]] explaining why it's not a payload
    * or an empty response.
    */
-  private[finch] final case class Failure(cause: Exception, meta: Meta) extends Output[Nothing] {
-    override protected def withMeta(meta: Meta): Output[Nothing] = copy(meta = meta)
-    override def value: Nothing = throw cause
+  private[finch] final case class Failure(
+      cause: Exception,
+      status: Status = Status.BadRequest,
+      charset: Option[Charset] = Option.empty,
+      headers: Map[String, String] = Map.empty[String, String],
+      cookies: Seq[Cookie] = Seq.empty[Cookie]) extends Output[Nothing] {
+
+    def value: Nothing = throw cause
+
+    protected def copy(
+        status: Status,
+        charset: Option[Charset],
+        headers: Map[String, String],
+        cookies: Seq[Cookie]): Output[Nothing] = Failure(cause, status, charset, headers, cookies)
   }
 
   /**
    * An empty [[Output]] that does not capture any payload.
    */
-  private[finch] final case class Empty(meta: Meta) extends Output[Nothing] {
-    override protected def withMeta(meta: Meta): Output[Nothing] = copy(meta = meta)
-    override def value: Nothing = throw new IllegalStateException("empty output")
+  private[finch] final case class Empty(
+      status: Status,
+      charset: Option[Charset] = Option.empty,
+      headers: Map[String, String] = Map.empty[String, String],
+      cookies: Seq[Cookie] = Seq.empty[Cookie]) extends Output[Nothing] {
+
+    def value: Nothing = throw new IllegalStateException("empty output")
+
+    protected def copy(
+        status: Status,
+        charset: Option[Charset],
+        headers: Map[String, String],
+        cookies: Seq[Cookie]): Output[Nothing] = Empty(status, charset, headers, cookies)
   }
 
-  implicit def outputEq[A](implicit A: Eq[A]): Eq[Output[A]] = Eq.instance {
-    case (Payload(av, am), Payload(bv, bm)) => A.eqv(av, bv) && am == bm
-    case (Failure(ac, am), Failure(bc, bm)) => ac == bc && am == bm
-    case (Empty(am), Empty(bm)) => am == bm
-    case (_, _) => false
-  }
+  implicit def outputEq[A]: Eq[Output[A]] = Eq.fromUniversalEquals
 
   implicit class OutputOps[A](val o: Output[A]) extends AnyVal {
 
@@ -175,9 +200,9 @@ object Output {
       tre: ToResponse.Aux[Exception, CT]
     ): Response = {
       val rep = o match {
-        case Output.Payload(v, m) => tr(v, m.charset.getOrElse(StandardCharsets.UTF_8))
-        case Output.Failure(x, m) => tre(x, m.charset.getOrElse(StandardCharsets.UTF_8))
-        case Output.Empty(_) => Response()
+        case p: Output.Payload[A] => tr(p.value, p.charset.getOrElse(StandardCharsets.UTF_8))
+        case f: Output.Failure => tre(f.cause, f.charset.getOrElse(StandardCharsets.UTF_8))
+        case e: Output.Empty => Response()
       }
 
       rep.status = o.status