Applying some technics I learned from @fwbrasil.

The biggest thing is to avoid closure allocations by embedding a `Funciton1` or `Function2` into an enclosing scope/object. This played really nicely in the `Endpoint` (see performance results bellow).

In addition to that optimization I also restructured the `Output` type a bit:

- Got rid of `Meta` and merged its members into each object
- Added `traverseFlatten` (a specialised and cheap version of `traverse().flatten`)
- Added `withHeaders` and `withCookies` that allow copying all things in one shot (no need to allocate an intermediate `Output` for each header).

In a nutshell
- `map*` operations are now up to **40% faster** (400 bytes fewer allocations on each call)
- `::` is now up to **5% faster** (120 bytes fewer allocations on each call)

Running time:

```
NEW:

MapBenchmark.map                             avgt   10   633.837 ±  14.613   ns/op
MapBenchmark.mapAsync                        avgt   10   624.249 ±  22.426   ns/op
MapBenchmark.mapOutput                       avgt   10   734.677 ±  30.215   ns/op
MapBenchmark.mapOutputAsync                  avgt   10   737.170 ±  29.808   ns/op
ProductBenchmark.bothMatched                 avgt   10  1175.716 ±  44.236   ns/op
ProductBenchmark.leftMatched                 avgt   10    26.510 ±   2.335   ns/op
ProductBenchmark.rightMatched                avgt   10     5.081 ±   0.112   ns/op

OLD:

MapBenchmark.map                             avgt   10   624.174 ±  28.621   ns/op
MapBenchmark.mapAsync                        avgt   10   647.369 ±  30.775   ns/op
MapBenchmark.mapOutput                       avgt   10  1221.053 ±  39.319   ns/op
MapBenchmark.mapOutputAsync                  avgt   10  1202.541 ±  44.432   ns/op
ProductBenchmark.bothMatched                 avgt   10  1224.278 ±  49.114   ns/op
ProductBenchmark.leftMatched                 avgt   10    29.856 ±   0.709   ns/op
ProductBenchmark.rightMatched                avgt   10     5.209 ±   0.112   ns/op
```

Allocations:

```
NEW: 

MapBenchmark.map:·gc.alloc.rate.norm                 avgt   10   664.000 ±   0.001    B/op
MapBenchmark.mapAsync:·gc.alloc.rate.norm            avgt   10   664.000 ±   0.001    B/op
MapBenchmark.mapOutput:·gc.alloc.rate.norm           avgt   10   832.000 ±   0.001    B/op
MapBenchmark.mapOutputAsync:·gc.alloc.rate.norm      avgt   10   832.000 ±   0.001    B/op
ProductBenchmark.bothMatched:·gc.alloc.rate.norm     avgt   10  1400.001 ±   0.001    B/op
ProductBenchmark.leftMatched:·gc.alloc.rate.norm     avgt   10    80.000 ±   0.001    B/op
ProductBenchmark.rightMatched:·gc.alloc.rate.norm    avgt   10    ≈ 10⁻⁶              B/op

OLD:

MapBenchmark.map:·gc.alloc.rate.norm                 avgt   10   696.000 ±   0.001    B/op
MapBenchmark.mapAsync:·gc.alloc.rate.norm            avgt   10   696.000 ±   0.001    B/op
MapBenchmark.mapOutput:·gc.alloc.rate.norm           avgt   10  1232.001 ±   0.001    B/op
MapBenchmark.mapOutputAsync:·gc.alloc.rate.norm      avgt   10  1232.001 ±   0.001    B/op
ProductBenchmark.bothMatched:·gc.alloc.rate.norm     avgt   10  1520.001 ±   0.001    B/op
ProductBenchmark.leftMatched:·gc.alloc.rate.norm     avgt   10   104.000 ±   0.001    B/op
ProductBenchmark.rightMatched:·gc.alloc.rate.norm    avgt   10    ≈ 10⁻⁶              B/op
```

/cc @travisbrown @rpless