@vkostyukov 2017-06-16
Thanks, @rpless! Yeah, these numbers are average time and allocations: less is better.